Move stores and actions
