+++ graylog2-web-interface/src/actions/sidecars/CollectorConfigurationsActions.js
