+++ graylog2-web-interface/src/actions/sidecars/CollectorsActions.js
