+++ graylog2-web-interface/src/actions/sidecars/SidecarsActions.js
