+++ graylog2-web-interface/src/actions/sidecars/SidecarsAdministrationActions.js
@@ -1,0 +1,8 @@
+import Reflux from 'reflux';
+
+const SidecarsAdministrationActions = Reflux.createActions({
+  list: { asyncResult: true },
+  setAction: { asyncResult: true },
+});
+
+export default SidecarsAdministrationActions;
