--- graylog2-web-interface/src/components/sidecars/administration/AdministrationActions.js
@@ -1,8 +1,0 @@
-import Reflux from 'reflux';
-
-const AdministrationActions = Reflux.createActions({
-  list: { asyncResult: true },
-  setAction: { asyncResult: true },
-});
-
-export default AdministrationActions;
