+++ graylog2-web-interface/src/injection/CombinedProvider.js
@@ -10,10 +10,8 @@ class CombinedProvider {
       Authentication: () => require('actions/authentication/AuthenticationActions'),
       ClusterTraffic: () => require('actions/cluster/ClusterTrafficActions'),
       CodecTypes: () => require('actions/codecs/CodecTypesActions'),
-      Administration: () => require('components/sidecars/administration/AdministrationActions'),
-      Collectors: () => require('components/sidecars/configurations/CollectorsActions'),
-      CollectorConfigurations: () => require('components/sidecars/configurations/CollectorConfigurationsActions'),
-      Sidecars: () => require('components/sidecars/sidecars/SidecarsActions'),
+      CollectorConfigurations: () => require('actions/sidecars/CollectorConfigurationsActions'),
+      Collectors: () => require('actions/sidecars/CollectorsActions'),
       Configuration: () => require('actions/configurations/ConfigurationActions'),
       ConfigurationBundles: () => require('actions/configuration-bundles/ConfigurationBundlesActions'),
       Dashboards: () => require('actions/dashboards/DashboardsActions'),
@@ -46,6 +44,8 @@ class CombinedProvider {
       SavedSearches: () => require('actions/search/SavedSearchesActions'),
       ServerAvailability: () => require('actions/sessions/ServerAvailabilityActions'),
       Session: () => require('actions/sessions/SessionActions'),
+      Sidecars: () => require('actions/sidecars/SidecarsActions'),
+      SidecarsAdministration: () => require('actions/sidecars/SidecarsAdministrationActions'),
       SingleNode: () => require('actions/nodes/SingleNodeActions'),
       Streams: () => require('actions/streams/StreamsActions'),
       SystemJobs: () => require('actions/systemjobs/SystemJobsActions'),
@@ -61,10 +61,8 @@ class CombinedProvider {
       ClusterOverview: () => require('stores/cluster/ClusterOverviewStore'),
       ClusterTraffic: () => require('stores/cluster/ClusterTrafficStore'),
       CodecTypes: () => require('stores/codecs/CodecTypesStore'),
-      Administration: () => require('components/sidecars/administration/AdministrationStore'),
-      Collectors: () => require('components/sidecars/configurations/CollectorsStore'),
-      CollectorConfigurations: () => require('components/sidecars/configurations/CollectorConfigurationsStore'),
-      Sidecars: () => require('components/sidecars/sidecars/SidecarsStore'),
+      CollectorConfigurations: () => require('stores/sidecars/CollectorConfigurationsStore'),
+      Collectors: () => require('stores/sidecars/CollectorsStore'),
       ConfigurationBundles: () => require('stores/configuration-bundles/ConfigurationBundlesStore'),
       Configurations: () => require('stores/configurations/ConfigurationsStore'),
       CurrentUser: () => require('stores/users/CurrentUserStore'),
@@ -114,6 +112,8 @@ class CombinedProvider {
       Search: () => require('stores/search/SearchStore'),
       ServerAvailability: () => require('stores/sessions/ServerAvailabilityStore'),
       Session: () => require('stores/sessions/SessionStore'),
+      Sidecars: () => require('stores/sidecars/SidecarsStore'),
+      SidecarsAdministration: () => require('stores/sidecars/SidecarsAdministrationStore'),
       SingleNode: () => require('stores/nodes/SingleNodeStore'),
       Sources: () => require('stores/sources/SourcesStore'),
       Startpage: () => require('stores/users/StartpageStore'),
