+++ graylog2-web-interface/src/stores/sidecars/CollectorConfigurationsStore.js
