+++ graylog2-web-interface/src/stores/sidecars/CollectorsStore.js
