+++ graylog2-web-interface/src/stores/sidecars/SidecarsStore.js
